Rename base.py for pytest discovery
