+++ testsuite/MDAnalysisTests/topology/test_dlpoly.py
@@ -27,7 +27,7 @@ from numpy.testing import (
 
 import MDAnalysis as mda
 
-from MDAnalysisTests.topology.base import ParserBase
+from MDAnalysisTests.topology.test_base import ParserBase
 from MDAnalysisTests.datafiles import (
     DLP_CONFIG,
     DLP_CONFIG_order,
