+++ testsuite/MDAnalysisTests/topology/test_xpdb.py
@@ -22,7 +22,7 @@
 from __future__ import absolute_import
 import MDAnalysis as mda
 
-from MDAnalysisTests.topology.base import ParserBase
+from MDAnalysisTests.topology.test_base import ParserBase
 from MDAnalysisTests.datafiles import (
     XPDB_small,
 )
